refactor: 集成项目资源发现到 project 模块
- 将 ProjectDiscovery 从 resource/discovery 移到 project 目录
- 在 project 初始化时自动扫描项目资源
- 更新工具描述强调资源发现价值
- ProjectArea 显示已发现的资源数量
- 修复资源扫描返回值处理

现在 project 工具会在初始化时自动扫描 .promptx/resource 目录，
发现并注册所有项目级角色、工具等资源。
