+++ packages/core/src/pouch/commands/ProjectCommand.js
@@ -4,7 +4,7 @@ const StateArea = require('../areas/common/StateArea')
 const { getGlobalResourceManager } = require('../../resource')
 const { COMMANDS, PACKAGE_NAMES } = require('~/constants')
 const RegistryData = require('../../resource/RegistryData')
-const ProjectDiscovery = require('../../resource/discovery/ProjectDiscovery')
+const ProjectDiscovery = require('../../project/ProjectDiscovery')
 const ProjectManager = require('~/project/ProjectManager')
 const { getGlobalProjectManager } = require('~/project/ProjectManager')
 const logger = require('@promptx/logger')
